pyproject.toml
[project]
name = "exrio"
version = "0.0.0-dev"
description = "Read and write OpenEXR files in Python with no system dependencies."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy<2",
]

[tool.uv]
cache-keys = [
    {file = "pyproject.toml"},
    {file = "Cargo.toml"},
    {file = "**/*.rs"},
]

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[dependency-groups]
dev = [
    "matplotlib>=3.9.4",
    "maturin>=1.7.8",
    "opencolorio>=2.4.1",
    "pillow>=11.0.0",
    "pytest>=8.3.4",
    "requests>=2.32.3",
    "scipy>=1.13.1",
]

[tool.maturin]
python-source = "src"
python-packages = ["exrio"]
features = ["pyo3/extension-module"]
module-name = "exrio._rust"
manifest-path = "rust/Cargo.toml"
